Split and refactor booking window config
