+++ server/routes/prisons/configuration/prisonConfigController.test.ts
@@ -13,7 +13,6 @@ let flashData: Record<string, string | FlashErrorMessage>
 const prisonService = createMockPrisonService()
 const sessionTemplateService = createMockSessionTemplateService()
 
-const allPrisons = TestData.prisons()
 const activePrison = TestData.prison()
 const inactivePrison = TestData.prison({ active: false })
 const prisonContactDetails = TestData.prisonContactDetails()
@@ -22,7 +21,6 @@ beforeEach(() => {
   flashData = {}
   flashProvider.mockImplementation(key => flashData[key])
 
-  prisonService.getAllPrisons.mockResolvedValue(allPrisons)
   prisonService.getPrison.mockResolvedValue(activePrison)
   prisonService.getPrisonContactDetails.mockResolvedValue(prisonContactDetails)
 
@@ -31,119 +29,6 @@ beforeEach(() => {
 
 afterEach(() => {
   jest.resetAllMocks()
-})
-
-describe('Prison booking window edit', () => {
-  const editBookingWindowUrl = `/prisons/HEI/configuration/booking-window/edit`
-
-  describe(`GET ${editBookingWindowUrl}`, () => {
-    it('should render edit booking window form', () => {
-      prisonService.getPrison.mockResolvedValue(activePrison)
-
-      return request(app)
-        .get(editBookingWindowUrl)
-        .expect(200)
-        .expect('Content-Type', /html/)
-        .expect(res => {
-          const $ = cheerio.load(res.text)
-          expect($('h1').text().trim()).toContain('Edit prison booking window')
-          expect($('#policyNoticeDaysMin').val()).toBe('2')
-          expect($('#policyNoticeDaysMax').val()).toBe('28')
-          expect($('[data-test="submit"]').text().trim()).toBe('Update')
-        })
-        .expect(() => {
-          expect(prisonService.getPrison).toHaveBeenCalledTimes(1)
-        })
-    })
-  })
-
-  describe(`POST  ${editBookingWindowUrl}`, () => {
-    prisonService.getPrison.mockResolvedValue(activePrison)
-
-    it('should send valid data to edit booking window and redirect to view template', () => {
-      const updatePrisonDto = TestData.updatePrisonDto({ policyNoticeDaysMin: 10, policyNoticeDaysMax: 20 })
-
-      return request(app)
-        .post(editBookingWindowUrl)
-        .send(`policyNoticeDaysMin=${updatePrisonDto.policyNoticeDaysMin}`)
-        .send(`policyNoticeDaysMax=${updatePrisonDto.policyNoticeDaysMax}`)
-        .expect(302)
-        .expect('Location', `/prisons/${activePrison.code}/configuration`)
-        .expect(() => {
-          expect(flashProvider.mock.calls.length).toBe(1)
-          expect(flashProvider).toHaveBeenCalledWith('message', 'Booking window updated')
-          expect(prisonService.updatePrisonDetails).toHaveBeenCalledWith('user1', activePrison.code, updatePrisonDto)
-        })
-    })
-
-    it('should set validation errors when min and max are below 0 form data', () => {
-      const updatePrisonDto = TestData.updatePrisonDto({ policyNoticeDaysMin: -1, policyNoticeDaysMax: -1 })
-
-      const expectedValidationErrors = [
-        expect.objectContaining({ path: 'policyNoticeDaysMin', msg: 'Enter a min booking window value of at least 1' }),
-        expect.objectContaining({ path: 'policyNoticeDaysMax', msg: 'Enter a max booking window value of at least 1' }),
-      ]
-
-      const expectedFormValues = updatePrisonDto
-
-      return request(app)
-        .post(editBookingWindowUrl)
-        .send(`policyNoticeDaysMin=${updatePrisonDto.policyNoticeDaysMin}`)
-        .send(`policyNoticeDaysMax=${updatePrisonDto.policyNoticeDaysMax}`)
-        .expect(302)
-        .expect('Location', `/prisons/HEI/configuration/booking-window/edit`)
-        .expect(() => {
-          expect(flashProvider.mock.calls.length).toBe(2)
-          expect(flashProvider).toHaveBeenCalledWith('errors', expect.arrayContaining(expectedValidationErrors))
-          expect(flashProvider).toHaveBeenCalledWith('formValues', expectedFormValues)
-          expect(prisonService.updatePrisonDetails).not.toHaveBeenCalled()
-        })
-    })
-
-    it('should set validation errors when min is greater than max', () => {
-      const updatePrisonDto = TestData.updatePrisonDto({ policyNoticeDaysMin: 10, policyNoticeDaysMax: 1 })
-
-      const expectedValidationErrors = [
-        expect.objectContaining({
-          path: 'policyNoticeDaysMin',
-          msg: 'Enter a Min window less than or equal to the Max',
-        }),
-      ]
-
-      const expectedFormValues = updatePrisonDto
-
-      return request(app)
-        .post(editBookingWindowUrl)
-        .send(`policyNoticeDaysMin=${updatePrisonDto.policyNoticeDaysMin}`)
-        .send(`policyNoticeDaysMax=${updatePrisonDto.policyNoticeDaysMax}`)
-        .expect(302)
-        .expect('Location', `/prisons/HEI/configuration/booking-window/edit`)
-        .expect(() => {
-          expect(flashProvider.mock.calls.length).toBe(2)
-          expect(flashProvider).toHaveBeenCalledWith('errors', expect.arrayContaining(expectedValidationErrors))
-          expect(flashProvider).toHaveBeenCalledWith('formValues', expectedFormValues)
-          expect(prisonService.updatePrisonDetails).not.toHaveBeenCalled()
-        })
-    })
-
-    it('should handle API errors by setting flash errors and redirecting to same page', () => {
-      const updatePrisonDto = TestData.updatePrisonDto({ policyNoticeDaysMin: 10, policyNoticeDaysMax: 20 })
-      prisonService.updatePrisonDetails.mockRejectedValue(new BadRequest('API error!'))
-
-      return request(app)
-        .post(editBookingWindowUrl)
-        .send(`policyNoticeDaysMin=${updatePrisonDto.policyNoticeDaysMin}`)
-        .send(`policyNoticeDaysMax=${updatePrisonDto.policyNoticeDaysMax}`)
-        .expect(302)
-        .expect('Location', `/prisons/${activePrison.code}/configuration/booking-window/edit`)
-        .expect(() => {
-          expect(prisonService.updatePrisonDetails).toHaveBeenCalledWith('user1', activePrison.code, updatePrisonDto)
-          expect(flashProvider.mock.calls.length).toBe(2)
-          expect(flashProvider).toHaveBeenCalledWith('errors', [{ msg: '400 API error!' }])
-          expect(flashProvider).toHaveBeenCalledWith('formValues', updatePrisonDto)
-        })
-    })
-  })
 })
 
 describe('Prison configuration', () => {
@@ -172,7 +57,8 @@ describe('Prison configuration', () => {
 
             expect($('h2').eq(0).text().trim()).toContain('Contact details')
             expect($('h2').eq(1).text().trim()).toContain('Change booking window')
-            expect($('h2').eq(2).text().trim()).toContain('Change prison status')
+            expect($('h2').eq(2).text().trim()).toContain('Visitor configuration')
+            expect($('h2').eq(3).text().trim()).toContain('Change prison status')
           })
       })
